requirements.win.txt
cnocr
adb-shell
opencv-python
py-machineid
pyinstaller
PyYAML
setuptools
dhash
nicegui
pywin32
pywebview
onnxruntime
parse